+++ tests/captchas-5-8.test.js
@@ -113,7 +113,8 @@ function stubAudio() {
     blip() {},
     buzz() {},
     holyPad() {},
-    stopMusic() {}
+    stopMusic() {},
+    startMusic() {}
   };
 }